CoCo 1 Restoration
I bought this on eBay in 2018 for < $100.

The original owner thought it would be a great idea and paint it white to look like the CoCo 2 and 3.

The paint job was pretty bad.  In fact, I could scrap it off with my fingernail.

So I have been going through the process of stripping the paint and painting it the original colors (as best as I can).
I stripped the original paint (mostly) and put five layers of white primer.  Between each layer, I also sanded it with 1000 grit sand paper (very lightly).   It's as smooth as an android's bottom.